Int width and height

+++ picture.go
@@ -84,7 +84,7 @@ func (w *WAD) GetPicture(name string) (*Picture, error) {
 	}
 
 	// Cache picture
-	w.Pictures[name] = &Picture{Width: float64(header.Width), Height: float64(header.Height), Columns: columns}
+	w.Pictures[name] = &Picture{Width: int(header.Width), Height: int(header.Height), Columns: columns}
 
 	// Return pic
 	return w.Pictures[name], nil
